Doc fix: The database setting doesn't actually allow string values

+++ hypothesis-python/scripts/basic-test.sh
@@ -33,7 +33,10 @@ pip uninstall -y redis fakeredis
 
 pip install typing_extensions
 $PYTEST tests/typing_extensions/
-pip uninstall -y typing_extensions
+if [ "$(python -c 'import sys; print(sys.version_info[:2] <= (3, 7))')" = "False" ] ; then
+  # Required by importlib_metadata backport, which we don't want to break
+  pip uninstall -y typing_extensions
+fi
 
 pip install ".[lark]"
 $PYTEST tests/lark/
